Turned on security

+++ controllers/patients.ts
@@ -1,8 +1,8 @@
 import { Router } from "express";
 import { Patient } from "models";
-// import { Roles } from "types";
-import { getPatient } from "utils";
-// import { ForbiddenError } from "utils/errors";
+import { Roles } from "types";
+import { getLoggedUser, getPatient } from "utils";
+import { ForbiddenError } from "utils/errors";
 
 export const patientsRouter = Router();
 
@@ -58,17 +58,17 @@ patientsRouter.get("/:id", async (req, res, next) => {
 
 patientsRouter.put("/:id", async (req, res, next) => {
   try {
-    // const { role: userRole, id: userId } = await getLoggedUser(req);
+    const { role: userRole, id: userId } = await getLoggedUser(req);
     const patientId = req.params.id;
     const payload = req.body;
 
     const oldPatient = await getPatient(patientId);
 
-    // if (
-    //   !(userRole === Roles.ADMIN || userRole === Roles.MODERATOR) &&
-    //   oldPatient.author.id !== userId
-    // )
-    //   throw new ForbiddenError();
+    if (
+      !(userRole === Roles.ADMIN || userRole === Roles.MODERATOR) &&
+      oldPatient.author.id !== userId
+    )
+      throw new ForbiddenError();
 
     const updatedPatient = await oldPatient.updateOne(payload);
 
@@ -80,15 +80,15 @@ patientsRouter.put("/:id", async (req, res, next) => {
 
 patientsRouter.delete("/:id", async (req, res, next) => {
   try {
-    // const { role: userRole, id: userId } = await getLoggedUser(req);
+    const { role: userRole, id: userId } = await getLoggedUser(req);
     const patientId = req.params.id;
 
     const patient = await getPatient(patientId);
-    // if (
-    //   !(userRole === Roles.ADMIN || userRole === Roles.MODERATOR) &&
-    //   patient.author.id !== userId
-    // )
-    //   throw new ForbiddenError();
+    if (
+      !(userRole === Roles.ADMIN || userRole === Roles.MODERATOR) &&
+      patient.author.id !== userId
+    )
+      throw new ForbiddenError();
 
     await Patient.findByIdAndDelete(patient.id);
